Delete `timer.ts`
